Use cpanm instead of cpan to install dependencies
Previously, when using cpan to install the required Perl modules, the
install would fail in the case where the required Perl modules depend on
XS modules and the local lib directory at ~/perl5 does no yet exist.  In
this case cpan fails to properly resolve the location of these
architecture-specific modules.  Switch to cpanm as it does not suffer
from the same issue.